Added resync time in RunSynchronousComman
If for some reason the surface bios has another time than the real time we could get some issues when automating a install with OOBE.
Example in OOBE we create a user with a password and also setup autologin. If bios time is wrong it updates the windows-time efter it created the user rendering that password to be expired. If the time differance is 42days that is.